requirements.txt
langchain
langchain-core

langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
sentence-transformers
huggingface-hub

python-dotenv
pydantic
pydantic[email]

numpy==1.26.4
pandas
scikit-learn

torch

streamlit

